Adding trybots for Periodic and PerCommit Housekeeper
Created to fix the feature request https://code.google.com/p/skia/issues/detail?id=1139: support tries on the housekeeping bot

(SkipBuildbotRuns)
Review URL: https://codereview.chromium.org/12380071

git-svn-id: http://skia.googlecode.com/svn/buildbot@7950 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ master/skia_master_scripts/housekeeping_periodic_factory.py
@@ -30,30 +30,31 @@ class HouseKeepingPeriodicFactory(skia_factory.SkiaFactory):
     if clobber:
       self.AddSlaveScript(script='clean.py', description='Clean')
 
-    sanitize_script_path = self.TargetPathJoin('tools',
-                                               'sanitize_source_files.py')
-    skia_trunk_svn_baseurl = '%s/%s' % (
-        SKIA_SVN_BASEURL.replace('http', 'https'), 'trunk')
-    # Run the sanitization script.
-    self._skia_cmd_obj.AddRunCommand(
-        command='python %s' % sanitize_script_path,
-        description='RunSanitization')
-    if self._do_upload_results:
-      merge_dir_path = self.TargetPathJoin(tempfile.gettempdir(),
-                                           'sanitize-merge')
-      # Cleanup the previous (if any) sanitize merge dir.
+    if not self._do_patch_step:  # Do not run the sanitizer if it is a try job.
+      sanitize_script_path = self.TargetPathJoin('tools',
+                                                 'sanitize_source_files.py')
+      skia_trunk_svn_baseurl = '%s/%s' % (
+          SKIA_SVN_BASEURL.replace('http', 'https'), 'trunk')
+      # Run the sanitization script.
       self._skia_cmd_obj.AddRunCommand(
-        command='rm -rf %s' % merge_dir_path, description='Cleanup')
-      # Upload sanitized files.
-      self._skia_cmd_obj.AddMergeIntoSvn(
-          source_dir_path='.',
-          dest_svn_url=skia_trunk_svn_baseurl,
-          merge_dir_path=merge_dir_path,
-          svn_username_file=self._skia_svn_username_file,
-          svn_password_file=self._skia_svn_password_file,
-          commit_message=WithProperties(
-              'Sanitizing source files in %s' % self._builder_name),
-          description='UploadSanitizedFiles')
+          command='python %s' % sanitize_script_path,
+          description='RunSanitization')
+      if self._do_upload_results:
+        merge_dir_path = self.TargetPathJoin(tempfile.gettempdir(),
+                                             'sanitize-merge')
+        # Cleanup the previous (if any) sanitize merge dir.
+        self._skia_cmd_obj.AddRunCommand(
+          command='rm -rf %s' % merge_dir_path, description='Cleanup')
+        # Upload sanitized files.
+        self._skia_cmd_obj.AddMergeIntoSvn(
+            source_dir_path='.',
+            dest_svn_url=skia_trunk_svn_baseurl,
+            merge_dir_path=merge_dir_path,
+            svn_username_file=self._skia_svn_username_file,
+            svn_password_file=self._skia_svn_password_file,
+            commit_message=WithProperties(
+                'Sanitizing source files in %s' % self._builder_name),
+            description='UploadSanitizedFiles')
 
     # pylint: disable=W0212
     clang_static_analyzer_script_path = self.TargetPathJoin(
@@ -66,21 +67,22 @@ class HouseKeepingPeriodicFactory(skia_factory.SkiaFactory):
     self.AddSlaveScript(script='check_gs_timestamps.py',
                         description='CheckGoogleStorageTimestamps')
 
-    # pylint: disable=W0212
-    disk_usage_script_path = self.TargetPathJoin(
-        self._skia_cmd_obj._local_slave_script_dir,
-        'check_compute_engine_disk_usage.sh')
-    self._skia_cmd_obj.AddRunCommand(
-        command=('SKIA_COMPUTE_ENGINE_HOSTNAME=%s PERSISTENT_DISK_NAME='
-                 '/home/default/skia-master %s'
-                 % (SKIA_PUBLIC_MASTER, disk_usage_script_path)),
-        description='CheckMasterDiskUsage')
-    self._skia_cmd_obj.AddRunCommand(
-        command=(WithProperties('SKIA_COMPUTE_ENGINE_HOSTNAME=%(slavename)s '
-                                'PERSISTENT_DISK_NAME='
-                                '/home/default/skia-slave ' + \
-                                disk_usage_script_path)),
-        description='CheckHousekeepingSlaveDiskUsage')
+    if not self._do_patch_step:  # Do not run the checkers if it is a try job.
+      # pylint: disable=W0212
+      disk_usage_script_path = self.TargetPathJoin(
+          self._skia_cmd_obj._local_slave_script_dir,
+          'check_compute_engine_disk_usage.sh')
+      self._skia_cmd_obj.AddRunCommand(
+          command=('SKIA_COMPUTE_ENGINE_HOSTNAME=%s PERSISTENT_DISK_NAME='
+                   '/home/default/skia-master %s'
+                   % (SKIA_PUBLIC_MASTER, disk_usage_script_path)),
+          description='CheckMasterDiskUsage')
+      self._skia_cmd_obj.AddRunCommand(
+          command=(WithProperties('SKIA_COMPUTE_ENGINE_HOSTNAME=%(slavename)s '
+                                  'PERSISTENT_DISK_NAME='
+                                  '/home/default/skia-slave ' + \
+                                  disk_usage_script_path)),
+          description='CheckHousekeepingSlaveDiskUsage')
 
     return self
